refactor: move metadata and lyrics editing logic to MetadataEditStateHolder
Refactors `PlayerViewModel` by delegating metadata editing, batch updates, and lyrics management to `MetadataEditStateHolder`. This reduces ViewModel complexity and centralizes media persistence logic.

- Relocated metadata-related data classes (`PendingMetadataEdit`, `PendingBatchMetadataEdit`, `PendingLyricsSave`) to `MetadataEditStateHolder`.
- Introduced `MetadataEditCallbacks` to provide the state holder with controlled access to ViewModel-owned state and UI events.
- Moved implementation of `editSongMetadata`, `saveBatchMetadata`, `batchEditGenre`, and `saveLyricsToFile` along with their permission handling logic.
- Transferred ownership of the `writePermissionRequest` flow to the state holder.
- Cleaned up `PlayerViewModel` dependencies, removing direct injections of `AlbumArtThemeDao` and `ImageCacheManager`.
- Updated visibility of internal queue helper functions to support cross-component state updates.
