+++ app/src/test/java/com/theveloper/pixelplay/presentation/viewmodel/PlayerViewModelTest.kt
@@ -251,7 +251,6 @@ class PlayerViewModelTest {
             mockUserPreferencesRepository,
             mockAiPreferencesRepository,
             mockThemePreferencesRepository,
-            mockAlbumArtThemeDao,
             mockSyncManager,
             mockDualPlayerEngine,
             mockAppShortcutManager,
